Fix Various Absolute Mouse Coordinate Conversion Issues
This change fixes various absolute mouse coordinate conversion issues. It fixes 1) the video clamping range which instead of from 0 to effectiveWidth (or effectiveHeight) should instead be from 0 to effectiveWidth - 1 (or effectiveHeight - 1), 2) the mouse coordinate size which instead of 32767 should instead be 32768 since the min value is 0x0000 and the max value is 0x7FFF (32767), and 3) a rounding operation which should instead of a truncation operation.